Implements validation for categories and projects when adding an activity.
Adds exception classes in nonova_exceptions.
Changes NovaCliConn execute, it returns the output from the process call
instead of directly printing it, so it can be used to get and parse the
result.
Prints the result return by backend.execute where it was called.
Adds validate_project and validate_category, and implements them in
the new_activity method.